Extract DownloadInfoReader
